// angular6-example/src/app/model/examinationType.model.ts
export class ExaminationType{
    ExaminationTypeId: number;
    ExamGroup: string;
    ExamName: string;
    MinMarks: number;
    MaxMarks: number;
    FeeLabel: string;
    Amount: number;
    UniversityId
    Active: number;
    lastupdateddt:Date;
    lastupdatedby: number;
}





 